Update property setting for SceneKit light inner/outer angles that used to need setting via key-value. Update target button to match NSControl "weak" type.

+++ SceneKitTest/Classes/AEGeometry.h
@@ -105,7 +105,7 @@ CG_INLINE CGFloat AENormalize360(CGFloat angle) {
 		normalizedAngle = 0.0;
 	}
 
-	return fabsf(normalizedAngle);
+	return fabs(normalizedAngle);
 }
 
 /* =============================================================================
